% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erosion3.R
\name{erosion3}
\alias{erosion3}
\title{Erosion}
\usage{
erosion3(win, r)
}
\arguments{
\item{win}{A window of class \code{\link{box3}} or any data accepted by \code{\link{as.box3}}.}

\item{r}{The erosion factor.}
}
\value{
A window of class \code{\link{box3}}.
}
\description{
Function for eroding a cuboid by some value r
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
