% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStrauss3.R
\name{fitStrauss3}
\alias{fitStrauss3}
\title{Fit Strauss model}
\usage{
fitStrauss3(X, R, nd, erosion.factor = max(R),
  dum.sim.method = "stratified")
}
\arguments{
\item{X}{Three dimensional point pattern of class \code{\link{pp3}} contained in a window of class \code{\link{box3}}.}

\item{R}{A vector of interaction radii}

\item{nd}{For dum.sim.method = "Poisson" nd^3 is the expected number of dummy points in window eroded by.
For dum.sim.method = "stratified" nd specifies the grid dimension (grid is on the eroded window).}

\item{erosion.factor}{A number defining how much the window shall be eroded in the border correction.
If \code{0} no correction will be done. The default is \code{max(R)}.}

\item{dum.sim.method}{A string indicating which model the dummy point should be simulated from.
Options \code{"Poisson"} and \code{"stratified"}}
}
\value{
A \code{\link{data.frame}} with \code{length(R)} row with variables \code{R}, \code{beta}, \code{gamma} and \code{logLik}
}
\description{
Function for fitting the Strauss model using logistic regression to 3D data.
}
\details{
The method used is only for Gibbs point process models on the exponential family form.
For the Strauss process this means that the interaction radius should be fixed.
When these requirements are satisfied, a logistic regression is fitted.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
