% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3.R
\name{grid3}
\alias{grid3}
\title{Cuboid grid of points}
\usage{
grid3(win, nx, ny, nz)
}
\arguments{
\item{win}{A window of class \code{\link{box3}} or any data accepted by \code{\link{as.box3}}.}

\item{nx}{Number of points in the x direction of the grid.}

\item{ny}{Number of points in the y direction of the grid.}

\item{nz}{Number of points in the z direction of the grid.}
}
\value{
A list consisting of x, y and z which are numeric vectors constituting the coordinates of the cuboid grid.
}
\description{
Generate a cuboid grid of points in three dimensions window
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
