% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRStrauss.R
\name{HRStrauss}
\alias{HRStrauss}
\title{Hastings ratio for the Strauss model}
\usage{
HRStrauss(x, x_prop, xi, h, beta, gamma, R, birth = TRUE)
}
\arguments{
\item{x}{A point pattern of class \code{\link{pp3}}.}

\item{x_prop}{A point proposed to be added to the point pattern.}

\item{xi}{A point in \code{x}.}

\item{h}{The hardcore parameter.}

\item{beta}{The intensity parameter.}

\item{gamma}{The interaction parameter.}

\item{R}{The interactionradius parameter.}

\item{birth}{A parameter indicating whether the death or birth Hasting ratio should be calculated.}
}
\value{
A numeric value. The Hastings ratio.
}
\description{
Function used for calculating the hastings ratio used for simulation with the Metropolis-Hastings algorithm, \code{\link{rStrauss}}.
}
\details{
Used internally.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
