% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J3est.R
\name{J3est}
\alias{J3est}
\title{J-function}
\usage{
J3est(X, ...)
}
\arguments{
\item{X}{An observed point pattern of class \code{\link{pp3}}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{\link{fv.object}} containing r the vector of radii at which the J-function is to be estimated.
}
\description{
Function for estimating the J-function from a three dimensional point pattern.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
