% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcyl3est.R
\name{Kcyl3est}
\alias{Kcyl3est}
\title{Cylindrical K-function}
\usage{
Kcyl3est(X, t, u, ..., rmax = NULL, nrval = 128, drop = TRUE)
}
\arguments{
\item{X}{A point pattern of class \code{\link{pp3}}.}

\item{t}{A numeric vector of heights.}

\item{u}{A numeric vector determining the direction of the cylinders.}

\item{...}{Ignored}

\item{rmax}{Optional. Maximum value of argument \code{r} for which \code{Kcyl3(r)} will be estimated.}

\item{nrval}{Optional. Number of values of r for which \code{Kcyl3(r)} will be estimated.
A large value of nrval is required to avoid discretisation effects.}

\item{drop}{A logical value indicating whether or not the object to be returned should be a list or an \code{\link{fv}} oject.
Only works when \code{length(t) == 1}.}
}
\value{
A list where each element is an object of class \code{\link{fv}} containing the \code{r} values,
theoretical Poisson values and the estimated function values.
Each list element corresponds to the different heights of the cylinders.
}
\description{
Function for estimating the cylindrical K-function from a three dimensional point pattern.
Note: the cylinder only extends in one direction of the reference point.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
