% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrat3.R
\name{rstrat3}
\alias{rstrat3}
\title{Stratified random sample of point}
\usage{
rstrat3(win, nx, ny = nx, nz = nx, k = 1)
}
\arguments{
\item{win}{A window of class \code{\link{box3}} or any data accepted by \code{\link{as.box3}}}

\item{nx}{The number of cuboids in the x direction, respectively.}

\item{ny}{The number of cuboids in the y direction, respectively.}

\item{nz}{The number of cuboids in the z direction, respectively.}

\item{k}{Number of points generated in each cuboid}
}
\value{
A \code{\link{data.frame}} with 3 columns and \code{nx*ny*nz*k} rows
}
\description{
Generates a stratified random point pattern i a window.
}
\details{
The window is divided into a number of cuboid tiles and \code{k} points are then uniformly sampled in each cuboid
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
