% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStrauss3.R
\name{rStrauss3}
\alias{rStrauss3}
\title{Simulate Strauss}
\usage{
rStrauss3(beta, gamma = 1, R = h, h = 0, win = box3(),
  multi.sim = FALSE, nit = 10000, trace = TRUE, Keep.an.eye = FALSE)
}
\arguments{
\item{beta}{The intensity parameter (a positive number).}

\item{gamma}{The interaction parameter (a number between 0 and 1, both inclusive).}

\item{R}{The interactionradius parameter (a number greater than \code{h}).}

\item{h}{The hardcore parameter (a non-negative number).}

\item{win}{A window in which the simulation should be contained. Should be of class \code{\link{box3}}
or any data accepted by \code{\link{as.box3}}.}

\item{multi.sim}{If \code{TRUE} a list of simulations will be returned. 
The "distance" between each simulated point pattern will be \code{1000}.}

\item{nit}{The number of iterations/point pattern to be simulated.}

\item{trace}{If \code{TRUE} the number of points, number of h-close neighbors, and number of R-close neighbors is printed.
The number of points and the number of R-close neighbors are the sufficient statistics.}

\item{Keep.an.eye}{If \code{TRUE} a number indicating the number of poit patterns so far simulated will be printed.}
}
\value{
A list of three dimensional point patterns
}
\description{
Function for simulating the Strauss (hardcore) model a number of times.
}
\details{
The initial point pattern is an empty Poisson point pattern.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
